replaced the welcome page with a home page
